rename : 프로젝트 패키지구조변경 - domain 영역 분할
