+++ src/main/java/com/sparta/i_am_delivery/domain/menu/entity/Menu.java
@@ -1,0 +1,34 @@
+package com.sparta.i_am_delivery.domain.menu.entity;
+
+import com.sparta.i_am_delivery.common.entity.TimeStamped;
+import com.sparta.i_am_delivery.domain.store.entity.Store;
+import jakarta.persistence.*;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Table(name = "menus")
+@NoArgsConstructor
+public class Menu extends TimeStamped {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "store_id", nullable = false)
+    private Store store;
+
+    @Column(nullable = false)
+    private String name;
+
+    private Double menuPrice;
+
+    @Builder
+    public Menu(Store store, String name, Double menuPrice) {
+        this.store = store;
+        this.name = name;
+        this.menuPrice = menuPrice;
+    }
+}
