+++ src/main/java/com/sparta/i_am_delivery/domain/user/entity/User.java
@@ -1,0 +1,39 @@
+package com.sparta.i_am_delivery.domain.user.entity;
+
+import com.sparta.i_am_delivery.common.entity.TimeStamped;
+import com.sparta.i_am_delivery.user.enums.UserType;
+import jakarta.persistence.*;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Table(name = "users")
+@NoArgsConstructor
+public class User extends TimeStamped {
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @Column(nullable = false, unique = true)
+  private String email;
+
+  @Column(nullable = false)
+  private String password;
+
+  @Column(nullable = false)
+  private String name;
+
+  @Column(nullable = false)
+  @Enumerated(value = EnumType.STRING)
+  private UserType type;
+
+  @Builder
+  public User(String email, String password, String name, UserType type) {
+    this.email = email;
+    this.password = password;
+    this.name = name;
+    this.type = type;
+  }
+}
