--- src/main/java/com/sparta/i_am_delivery/like/entity/Like.java
@@ -1,33 +1,0 @@
-package com.sparta.i_am_delivery.like.entity;
-
-import com.sparta.i_am_delivery.common.entity.TimeStamped;
-import com.sparta.i_am_delivery.user.entity.User;
-import com.sparta.i_am_delivery.store.entity.Store;
-import jakarta.persistence.*;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Table(name = "likes")
-@NoArgsConstructor
-public class Like extends TimeStamped {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "store_id", nullable = false)
-    private Store store;
-
-    @Builder
-    public Like(User user, Store store) {
-        this.user = user;
-        this.store = store;
-    }
-}
