--- src/main/java/com/sparta/i_am_delivery/order/entity/enums/OrderStatus.java
@@ -1,9 +1,0 @@
-package com.sparta.i_am_delivery.order.entity.enums;
-
-public enum OrderStatus {
-    RECEIVED,    // 접수중
-    PREPARING,   // 조리중
-    DELIVERING,  // 배달중
-    COMPLETED,   // 배달완료
-    CANCELED     // 주문취소
-}
