--- src/main/java/com/sparta/i_am_delivery/review/entity/Review.java
@@ -1,40 +1,0 @@
-package com.sparta.i_am_delivery.review.entity;
-
-import com.sparta.i_am_delivery.common.entity.TimeStamped;
-import com.sparta.i_am_delivery.user.entity.User;
-import com.sparta.i_am_delivery.store.entity.Store;
-import jakarta.persistence.*;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Table(name = "reviews")
-@NoArgsConstructor
-public class Review extends TimeStamped {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "store_id", nullable = false)
-    private Store store;
-
-    @Column(nullable = false)
-    private String comment;
-
-    private Double star;
-
-    @Builder
-    public Review(User user, Store store, String comment, Double star) {
-        this.user = user;
-        this.store = store;
-        this.comment = comment;
-        this.star = star;
-    }
-}
