--- src/main/java/com/sparta/i_am_delivery/store/entity/Store.java
@@ -1,38 +1,0 @@
-package com.sparta.i_am_delivery.store.entity;
-
-import com.sparta.i_am_delivery.common.entity.TimeStamped;
-import com.sparta.i_am_delivery.user.entity.User;
-import jakarta.persistence.*;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Table(name = "stores")
-@NoArgsConstructor
-public class Store extends TimeStamped {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User owner;
-
-    @Column(nullable = false)
-    private String name;
-
-    private String openTime;
-    private String closeTime;
-    private Double minimumPrice;
-
-    @Builder
-    public Store(User owner, String name, String openTime, String closeTime, Double minimumPrice) {
-        this.owner = owner;
-        this.name = name;
-        this.openTime = openTime;
-        this.closeTime = closeTime;
-        this.minimumPrice = minimumPrice;
-    }
-}
