+++ src/main/java/com/sparta/i_am_delivery/user/service/UserService.java
@@ -1,14 +1,13 @@
 package com.sparta.i_am_delivery.user.service;
 
-import com.sparta.i_am_delivery.common.config.jwt.JwtHelper;
 import com.sparta.i_am_delivery.common.config.security.PasswordEncoder;
 import com.sparta.i_am_delivery.common.exception.CustomException;
 import com.sparta.i_am_delivery.common.exception.ErrorCode;
-import com.sparta.i_am_delivery.user.dto.UserSingUpRequestDto;
-import com.sparta.i_am_delivery.user.dto.UserSingUpResponseDto;
-import com.sparta.i_am_delivery.user.entity.User;
+import com.sparta.i_am_delivery.user.dto.request.UserSignUpRequestDto;
+import com.sparta.i_am_delivery.user.dto.response.UserSignUpResponseDto;
+import com.sparta.i_am_delivery.domain.user.entity.User;
 import com.sparta.i_am_delivery.user.enums.UserType;
-import com.sparta.i_am_delivery.user.repository.UserRepository;
+import com.sparta.i_am_delivery.domain.user.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
@@ -21,17 +20,17 @@ public class UserService {
   private final PasswordEncoder passwordEncoder;
 
   @Transactional
-  public UserSingUpResponseDto singUp(UserSingUpRequestDto userSingupRequestDto) {
-    this.isDuplicateUser(userSingupRequestDto.getEmail(), userSingupRequestDto.getName());
+  public UserSignUpResponseDto signUp(UserSignUpRequestDto userSingUpRequestDto) {
+    this.isValidateUserUniqueness(userSingUpRequestDto.getEmail(), userSingUpRequestDto.getName());
     User user =
         User.builder()
-            .email(userSingupRequestDto.getEmail())
-            .password(passwordEncoder.encode(userSingupRequestDto.getPassword()))
-            .name(userSingupRequestDto.getName())
-            .type(UserType.valueOf(userSingupRequestDto.getType()))
+            .email(userSingUpRequestDto.getEmail())
+            .password(passwordEncoder.encode(userSingUpRequestDto.getPassword()))
+            .name(userSingUpRequestDto.getName())
+            .type(UserType.valueOf(userSingUpRequestDto.getType()))
             .build();
     userRepository.save(user);
-    return UserSingUpResponseDto.builder()
+    return UserSignUpResponseDto.builder()
         .id(user.getId())
         .email(user.getEmail())
         .name(user.getName())
@@ -39,7 +38,7 @@ public class UserService {
         .build();
   }
 
-  private void isDuplicateUser(String email, String name) {
+  private void isValidateUserUniqueness(String email, String name) {
     Optional<User> foundUser = userRepository.findByEmailOrName(email, name);
     if (foundUser.isPresent()) {
       User existingUser = foundUser.get();
