+++ src/test/java/com/sparta/i_am_delivery/user/service/UserServiceTest.java
@@ -6,11 +6,11 @@ import static org.mockito.Mockito.*;
 import com.sparta.i_am_delivery.common.config.security.PasswordEncoder;
 import com.sparta.i_am_delivery.common.exception.CustomException;
 import com.sparta.i_am_delivery.common.exception.ErrorCode;
-import com.sparta.i_am_delivery.user.dto.UserSingUpRequestDto;
-import com.sparta.i_am_delivery.user.dto.UserSingUpResponseDto;
-import com.sparta.i_am_delivery.user.entity.User;
+import com.sparta.i_am_delivery.user.dto.request.UserSignUpRequestDto;
+import com.sparta.i_am_delivery.user.dto.response.UserSignUpResponseDto;
+import com.sparta.i_am_delivery.domain.user.entity.User;
 import com.sparta.i_am_delivery.user.enums.UserType;
-import com.sparta.i_am_delivery.user.repository.UserRepository;
+import com.sparta.i_am_delivery.domain.user.repository.UserRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -28,10 +28,10 @@ class UserServiceTest {
 
   @Test
   @DisplayName("회원가입 성공")
-  void singUpSuccessTest() {
+  void signUpSuccessTest() {
     // given
-    UserSingUpRequestDto req =
-        UserSingUpRequestDto.builder()
+    UserSignUpRequestDto req =
+        UserSignUpRequestDto.builder()
             .email("aaa@gmail.com")
             .password("q1!w2@e3#r4$")
             .name("김사장")
@@ -47,7 +47,7 @@ class UserServiceTest {
     when(userRepository.save(any(User.class))).thenReturn(user);
     when(passwordEncoder.encode(req.getPassword())).thenReturn("encodedPassword");
     // when
-    UserSingUpResponseDto responseDto = userService.singUp(req);
+    UserSignUpResponseDto responseDto = userService.signUp(req);
     // then
     assertNotNull(responseDto);
     assertEquals(user.getEmail(), responseDto.getEmail());
@@ -60,8 +60,8 @@ class UserServiceTest {
   public void testSignUpWithDuplicateEmail() {
     // Given
     String duplicateEmail = "test@example.com";
-    UserSingUpRequestDto req =
-            UserSingUpRequestDto.builder()
+    UserSignUpRequestDto req =
+            UserSignUpRequestDto.builder()
                     .email(duplicateEmail)
                     .password("q1!w2@e3#r4$")
                     .name("김사장")
@@ -74,7 +74,7 @@ class UserServiceTest {
 
     // When & Then
     CustomException exception = assertThrows(CustomException.class, () -> {
-      userService.singUp(req);
+      userService.signUp(req);
     });
 
     assertEquals(ErrorCode.EMAIL_DUPLICATED, exception.getErrorCode());
@@ -85,8 +85,8 @@ class UserServiceTest {
   public void testSignUpWithDuplicateName() {
     // Given
     String duplicateName = "duplicateName";
-    UserSingUpRequestDto req =
-            UserSingUpRequestDto.builder()
+    UserSignUpRequestDto req =
+            UserSignUpRequestDto.builder()
                     .email("unique@example.com")
                     .password("q1!w2@e3#r4$")
                     .name(duplicateName)
@@ -99,7 +99,7 @@ class UserServiceTest {
 
     // When & Then
     CustomException exception = assertThrows(CustomException.class, () -> {
-      userService.singUp(req);
+      userService.signUp(req);
     });
 
     assertEquals(ErrorCode.USERNAME_DUPLICATED, exception.getErrorCode());
